escape character problems
So, when u take the argv the escape characters are separated into different indexes, so i am still having trouble identifying the escape characters in the string. So, if the input is hello\nworld, the backslash is at index 5 and the n is at index 6. Also, C doesn't accept '\' as a character which is why it is hard for me to find the escape characters.