Minor adjustment in call to split
i was getting an error when compiling that the call to split had too few arguments. There was no file name in the arguments like there should be based on the split function in tokenizer.c. I believe the name is in the variable "dir" so i fixed that, but I may be mistaken.