Updated Tokenizer to sort list of files
Changed "node" struct to "record" struct. Also, two functions were added. updateRecordList() looks to see if the current file is already in the list for this token. If not, it adds it to the end. If so, it increments the frequency and calls resortRecordList(). This function re-sorts the record list as necessary. I haven't tested it yet bc i haven't saved the test case files/directories to test with. But it does compile and I did a thorough test on paper with drawings for many different cases and I think it should work.